Implement a type mapping container
Most of the prototypes that I have done so far including some variation
of this structure. It maps from types to an arbitrary value while
recognizing subclass relationships. For example, if it contains an entry
for `int`, then it should use it as a mapping for `bool`. However, if it
has an entry for `bool` as well, then it should prefer the explicit
mapping.