Refactor rule processing
Since we only support URLSpec routes that point at RequestHandler
instances, we can require this much earlier in the processing.
Specifically, I moved the checks into OpenAPIApplication._process_rule
and changed the call signature of add_operation. I'm on the fence
about removing the explicit check in add_operation since it is
(technically) a public method. Essentially, the guard is now a type
checking annotation instead of a runtime check... acceptable?