Refactor coffee example into separate files.
I'm not completely sure whether this is better or worse but a single
file was getting pretty unwieldy. Instead of following the MVC
structure, I split to code along user lines. I kept models and placed
shared items into the "shared" module.

The plan is to add the barista endpoints in the "barista" module
alongside customer.