Remove the ability to subclass api.Body.
I never liked how this could muddy the object model. I also
found some cases where the the api.Body.__init__() method was
being skipped for some reason. I suspect that it is an edge
case in my understanding of Pydantic deserialization. In any
case, the feature is not worth supporting IMO. There is no need
to have another way to tag parameters.